linux-parallella: Rework recipe
Remove use of any linux-xlnx or meta-xilinx includes.

Setup recipe to use the "multi_v7_defconfig" as a base, and add small
config defaults on top (for epiphany, board and general). Instead of
using the linux-yocto configuration tools simply join the fragments
together. This removes the need for the large kernelConfigAdditions.cfg
fragment, and relies on the kernel source to enable default Zynq
support.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>