Port: ARM: dts: zynq: parallella: Fix up reserved-memory and memory nodes
Ported Nathan Rossi reserved-memory fix:

parallella-linux.git: 18c456a7ff58ff19b75180da412dd6ca0c200757

Setup the memory node so that it covers all the available memory. This
adds the chunk back that is specifically used for the elink as the
kernel will handle reserving it.

Fix the reserved-memory node for emem so that it has the 'no-map'
property as the property 'no-map' is not inherited from the parent node.
Also fix up the node name (missing a 0) to match the reg property.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>